# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Passport local proof of concept",
    "version": "0.1.0",
    "description": "Experimental Ed25519/JCS agent authority protocol. Acme issues capabilities; Globex makes the final decision. See docs/protocol.md for signing rules and docs/threat-model.md for trusted state and outage semantics."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Registry"
    },
    {
      "url": "http://127.0.0.1:8081",
      "description": "Globex"
    },
    {
      "url": "http://127.0.0.1:8082",
      "description": "Acme authority"
    },
    {
      "url": "http://127.0.0.1:8083",
      "description": "Independent Go verifier"
    }
  ],
  "paths": {
...
    "/v0/receipts": {
      "post": {
        "operationId": "recordAuthorizationReceipt",
        "summary": "recordAuthorizationReceipt",
        "servers": [
          {
            "url": "http://127.0.0.1:8080"
          }
        ],
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignedReceipt"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Created"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Administrator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Verification denied; may include a signed denial receipt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "Immutable ID conflict or concurrent update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "receipts",
        "summary": "Recent public authorization receipts (Globex serves execution/denial receipts at the same path on port 8081)",
        "servers": [
          {
            "url": "http://127.0.0.1:8080"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignedReceipt"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Administrator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Verification denied; may include a signed denial receipt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "Immutable ID conflict or concurrent update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v0/capabilities": {
      "post": {
        "operationId": "authorize",
        "summary": "Acme evaluates Cedar and signs an exact-request capability; runtime proof is required",
        "servers": [
          {
            "url": "http://127.0.0.1:8082"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationResult"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Administrator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Verification denied; may include a signed denial receipt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "Immutable ID conflict or concurrent update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v0/identity": {
      "post": {
        "operationId": "identity",
        "summary": "Acme verifies runtime identity without issuing authority",
        "servers": [
          {
            "url": "http://127.0.0.1:8082"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Administrator credential required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Verification denied; may include a signed denial receipt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "Immutable ID conflict or concurrent update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Random per-organization local administrator credential; organization creation uses a separate bootstrap credential. Signed mutations are also required."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "invalid_request",
              "invalid_signature",
              "expired",
              "revoked",
              "replay",
              "policy_mismatch",
              "untrusted_issuer",
              "denied",
              "request_mismatch",
              "invalid_proof",
              "invalid_nonce",
              "stale_state",
              "unavailable",
              "unauthorized",
              "conflict",
              "not_found",
              "feature_disabled"
            ]
          }
        }
      },
...
      "Created": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          }
        }
      },
...
      "Delegation": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "org_id",
          "agent_id",
          "runtime_key_id",
          "actions",
          "resources",
          "audiences",
          "not_before",
          "expires_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "org_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "agent_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "runtime_key_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
            },
            "minItems": 1,
            "maxItems": 32
          },
          "resources": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
            },
            "minItems": 1,
            "maxItems": 32
          },
          "audiences": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
            },
            "minItems": 1,
            "maxItems": 32
          },
          "not_before": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "expires_at": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          }
        }
      },
...
      "Request": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "request_id",
          "issuer_org",
          "agent_id",
          "runtime_key_id",
          "delegation_id",
          "action",
          "resource",
          "audience",
          "recipient",
          "amount_minor_units",
          "currency",
          "asset",
          "network",
          "method",
          "path",
          "body",
          "nonce",
          "issued_at",
          "expires_at",
          "want_proof"
        ],
        "properties": {
          "request_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "issuer_org": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "agent_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "runtime_key_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "delegation_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "action": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "resource": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "audience": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "recipient": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "amount_minor_units": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4294967295
          },
          "currency": {
            "type": "string",
            "const": "USD"
          },
          "asset": {
            "type": "string",
            "const": "iso4217:USD"
          },
          "network": {
            "type": "string",
            "const": "simulation"
          },
          "method": {
            "type": "string",
            "const": "POST"
          },
          "path": {
            "type": "string",
            "const": "/v0/purchases"
          },
          "body": {
            "type": "object",
            "additionalProperties": true
          },
          "nonce": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$",
            "minLength": 32
          },
          "issued_at": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "expires_at": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "want_proof": {
            "type": "boolean"
          }
        }
      },
      "Capability": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "capability_id",
          "issuer_org",
          "agent_id",
          "runtime_key_id",
          "delegation_id",
          "action",
          "resource",
          "audience",
          "recipient",
          "amount_minor_units",
          "currency",
          "asset",
          "network",
          "request_digest",
          "nonce",
          "issued_at",
          "expires_at",
          "policy_version",
          "policy_commitment",
          "proof_program"
        ],
        "properties": {
          "capability_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "issuer_org": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "agent_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "runtime_key_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "delegation_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "action": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "resource": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "audience": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "recipient": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "amount_minor_units": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4294967295
          },
          "currency": {
            "type": "string",
            "const": "USD"
          },
          "asset": {
            "type": "string",
            "const": "iso4217:USD"
          },
          "network": {
            "type": "string",
            "const": "simulation"
          },
          "request_digest": {
            "type": "string"
          },
          "nonce": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "issued_at": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "expires_at": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "policy_version": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "policy_commitment": {
            "type": "string"
          },
          "proof_program": {
            "type": "string"
          }
        }
      },
...
      "Receipt": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "receipt_type",
          "org_id",
          "decision",
          "code",
          "request_id",
          "capability_id",
          "verifier",
          "issued_at",
          "policy_version",
          "request_digest",
          "evidence_digest",
          "trace"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "receipt_type": {
            "type": "string",
            "enum": [
              "authorization",
              "execution",
              "denial"
            ]
          },
          "org_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "decision": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "capability_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "verifier": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "issued_at": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "policy_version": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "request_digest": {
            "type": "string"
          },
          "evidence_digest": {
            "type": "string"
          },
          "trace": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Proof": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "program",
          "proof"
        ],
        "properties": {
          "program": {
            "type": "string",
            "const": "passport.hidden-limit.bp5.ristretto.32x2.v1"
          },
          "proof": {
            "type": "string",
            "pattern": "^[a-f0-9]+$",
            "maxLength": 8192
          }
        }
      },
...
      "SignedDelegation": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "profile",
          "kind",
          "key_id",
          "payload",
          "signature"
        ],
        "properties": {
          "profile": {
            "type": "string",
            "const": "passport.v0.ed25519-jcs"
          },
          "kind": {
            "type": "string",
            "const": "delegation"
          },
          "key_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "payload": {
            "$ref": "#/components/schemas/Delegation"
          },
          "signature": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{86}$"
          }
        }
      },
...
      "SignedRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "profile",
          "kind",
          "key_id",
          "payload",
          "signature"
        ],
        "properties": {
          "profile": {
            "type": "string",
            "const": "passport.v0.ed25519-jcs"
          },
          "kind": {
            "type": "string",
            "const": "request"
          },
          "key_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "payload": {
            "$ref": "#/components/schemas/Request"
          },
          "signature": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{86}$"
          }
        }
      },
      "SignedCapability": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "profile",
          "kind",
          "key_id",
          "payload",
          "signature"
        ],
        "properties": {
          "profile": {
            "type": "string",
            "const": "passport.v0.ed25519-jcs"
          },
          "kind": {
            "type": "string",
            "const": "capability"
          },
          "key_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "payload": {
            "$ref": "#/components/schemas/Capability"
          },
          "signature": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{86}$"
          }
        }
      },
...
      "SignedReceipt": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "profile",
          "kind",
          "key_id",
          "payload",
          "signature"
        ],
        "properties": {
          "profile": {
            "type": "string",
            "const": "passport.v0.ed25519-jcs"
          },
          "kind": {
            "type": "string",
            "const": "receipt"
          },
          "key_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "payload": {
            "$ref": "#/components/schemas/Receipt"
          },
          "signature": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{86}$"
          }
        }
      },
...
      "AuthorizationInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "request",
          "delegation"
        ],
        "properties": {
          "request": {
            "$ref": "#/components/schemas/SignedRequest"
          },
          "delegation": {
            "$ref": "#/components/schemas/SignedDelegation"
          }
        }
      },
      "AuthorizationResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "capability",
          "authorization_receipt"
        ],
        "properties": {
          "capability": {
            "$ref": "#/components/schemas/SignedCapability"
          },
          "proof": {
            "$ref": "#/components/schemas/Proof"
          },
          "authorization_receipt": {
            "$ref": "#/components/schemas/SignedReceipt"
          }
        }
      },
...
      "Identity": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "verified",
          "agent_id",
          "runtime_key_id",
          "checked_at"
        ],
        "properties": {
          "verified": {
            "type": "boolean"
          },
          "agent_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "runtime_key_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9:._-]{0,127}$"
          },
          "checked_at": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          }
        }
      },
...
    }
  }
}